#660 - store - split user module
