+++ core/store/modules/user/actions.js
@@ -1,94 +1,14 @@
-import * as types from '../mutation-types'
-import config from 'config'
 import EventBus from 'core/plugins/event-bus'
-import { ValidationError } from 'core/lib/exceptions'
-import store from '../'
-import i18n from 'core/lib/i18n'
+import * as types from '../../mutation-types'
+import config from 'config'
+import store from '../../'
 const Ajv = require('ajv') // json validator
+import { ValidationError } from 'core/lib/exceptions'
+import i18n from 'core/lib/i18n'
 
-EventBus.$on('user-after-update', (event) => {
-  if (event.resultCode === 200) {
-    EventBus.$emit('notification', {
-      type: 'success',
-      message: i18n.t('Account data has successfully been updated'),
-      action1: { label: 'OK', action: 'close' }
-    })
-    store.dispatch('user/refreshCurrentUser', event.result)
-  }
-})
-
-EventBus.$on('session-after-started', (event) => { // example stock check callback
-  console.log('Loading user profile')
-  store.dispatch('user/me', { refresh: navigator.onLine }, { root: true }).then((us) => {
-  })
-})
-
-// After order has been placed fill in missing address information in user's profile
-EventBus.$on('order-after-placed', (order) => {
-  if (store.getters['user/isLoggedIn']) {
-    let currentUser = store.state.user.current
-    let hasShippingAddress = currentUser.hasOwnProperty('default_shipping')
-    let hasBillingAddress = currentUser.hasOwnProperty('default_billing')
-    if (!(hasShippingAddress && hasBillingAddress)) {
-      let customer = Object.assign({}, currentUser)
-      if (!hasShippingAddress) {
-        let shippingAddress = order.order.addressInformation.shippingAddress
-        customer.addresses.push({
-          firstname: shippingAddress.firstname,
-          lastname: shippingAddress.lastname,
-          street: [shippingAddress.street[0], shippingAddress.street[1]],
-          city: shippingAddress.city,
-          ...(shippingAddress.region ? { region: { region: shippingAddress.region } } : {}),
-          country_id: shippingAddress.country_id,
-          postcode: shippingAddress.postcode,
-          ...(shippingAddress.telephone ? { telephone: shippingAddress.telephone } : {}),
-          default_shipping: true
-        })
-      }
-      if (!hasBillingAddress) {
-        let billingAddress = order.order.addressInformation.billingAddress
-        let hasCompany = billingAddress.company
-        if (hasCompany) {
-          customer.addresses.push({
-            firstname: billingAddress.firstname,
-            lastname: billingAddress.lastname,
-            street: [billingAddress.street[0], billingAddress.street[1]],
-            city: billingAddress.city,
-            ...(billingAddress.region ? { region: { region: billingAddress.region } } : {}),
-            country_id: billingAddress.country_id,
-            postcode: billingAddress.postcode,
-            company: billingAddress.company,
-            vat_id: billingAddress.vat_id,
-            default_billing: true
-          })
-        }
-      }
-
-      store.dispatch('user/update', { customer: customer })
-    }
-  }
-})
-
-// initial state
-const state = {
-  token: '',
-  current: null,
-  session_started: new Date(),
-  newsletter: null
-}
-
-const getters = {
-  isLoggedIn (state) {
-    return state.current !== null
-  }
-}
-
-// actions
-const actions = {
-
+export default {
   startSession (context) {
     context.commit(types.USER_START_SESSION)
-
     const cache = global.db.usersCollection
     cache.getItem('current-token', (err, res) => {
       if (err) {
@@ -114,7 +34,6 @@ const actions = {
       }
     })
   },
-
   /**
    * Send password reset link for specific e-mail
    */
@@ -139,7 +58,6 @@ const actions = {
       return response
     })
   },
-
   /**
    * Login user and return user profile and current token
    */
@@ -161,7 +79,6 @@ const actions = {
         return resp
       })
   },
-
   /**
    * Login user and return user profile and current token
    */
@@ -270,15 +187,14 @@ const actions = {
       })
     }
   },
-
   refreshCurrentUser (context, userData) {
     context.commit(types.USER_INFO_LOADED, userData)
   },
-
   /**
    * Change user password
    */
   changePassword (context, passwordData) {
+    console.log(context)
     return fetch(config.users.endpoint + '/changePassword?token=' + context.state.token,
       {
         method: 'POST',
@@ -308,7 +224,6 @@ const actions = {
         }
       })
   },
-
   /**
    * Logout user
    */
@@ -321,7 +236,6 @@ const actions = {
       action1: { label: 'OK', action: 'close' }
     })
   },
-
   /**
    * Save user's newsletter preferences
    */
@@ -333,33 +247,4 @@ const actions = {
       action1: { label: 'OK', action: 'close' }
     })
   }
-}
-
-// mutations
-const mutations = {
-  [types.USER_TOKEN_CHANGED] (state, newToken) {
-    state.token = newToken
-  },
-  [types.USER_START_SESSION] (state) {
-    state.session_started = new Date()
-  },
-  [types.USER_INFO_LOADED] (state, currentUser) {
-    state.current = currentUser
-  },
-  [types.USER_END_SESSION] (state) {
-    state.token = ''
-    state.current = null
-    state.session_started = null
-  },
-  [types.USER_UPDATE_PREFERENCES] (state, newsletterPreferences) {
-    state.newsletter = newsletterPreferences
-  }
-}
-
-export default {
-  namespaced: true,
-  state,
-  getters,
-  actions,
-  mutations
 }
