+++ core/store/modules/user/getters.js
@@ -1,0 +1,5 @@
+export default {
+  isLoggedIn (state) {
+    return state.current !== null
+  }
+}
